% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDOE.R
\name{RDOE}
\alias{RDOE}
\title{Design of Experiments Module}
\usage{
RDOE()
}
\value{
RDOE is a graphic interface
}
\description{
Module of Design of Experiments 'RDOE' is a user – oriented interface that integrates the functions from the book "Design and Analysis of Experiments with R" (Lawson, 2014 <ISBN: 9781439868133>) to fulfill the requirements of an experimental research considering a completely randomized one factor statistical design covering descriptive statistics, basic plots, anova, multiple comparisons tests and diagnostics. It will consider more complicate statistical designs in the future.

You can learn more about this package at:
http://www.uv.mx/personal/nehuerta/rdoe/
}
\details{
RDOE is a package with a graphical interface dedicated to perform of Design of Experiments in R.

Note: RDOE is free software and comes with ABSOLUTELY NO WARRANTY.
}
\examples{
\dontrun{
##Install package
library(RDOE)
##Call the package
RDOE()
}

}
\references{
Lawson, J. (2014). Design and Analysis of Experiments with R.
CRC Press. ISBN: 9781439868133

Lawson, J. (2016). daewr: Design and Analysis of Experiments with R.
R package version 1.1-7. https://CRAN.R-project.org/package=daewr

Verzani, J. Based on the iwidgets code of Simon Urbanek, suggestions
by Simon Urbanek, Philippe Grosjean and Michael Lawrence (2014).
gWidgets: gWidgets API for building toolkit-independent, interactive GUIs.
R package version 0.0-54. https://CRAN.R-project.org/package=gWidgets
}
\author{
Eduardo Castaño-Tostado, Nery Sofia Huerta-Pacheco, Victor Manuel Aguirre-Torres
}
